refactor: restructure and update imports in `EditUserService` and related components
- Moved `DTO`, `Command`, and `Validators` to appropriate `Application` and `Domain` subdirectories.
- Adjusted import paths across affected files.
- Added `NegativePoints` error to `ErrorCatalog`.
